+++ packages/worker/src/app/router.tsx
@@ -5,7 +5,7 @@ import { Outbound } from '@/pages/outbound/Outbound';
 import { Camera } from '@/pages/camera/Camera';
 import Chat from '@/pages/chat/Chat';
 import Task from '@/pages/tasklist/Task';
-import {Inbound} from '../pages/inbound/Inbound';
+import {Inbound} from '../pages/inbound1/Inbound';
 
 
 const router = createBrowserRouter([
