+++ packages/worker/src/pages/inbound1/pages/LocationScanPage.tsx
