+++ packages/worker/src/pages/inbound1/ui/InspectionItem.tsx
