+++ packages/worker/src/pages/inbound1/ui/InspectionList.tsx
